% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{as_rcdf}
\alias{as_rcdf}
\title{Convert to \code{rcdf} class}
\usage{
as_rcdf(data)
}
\arguments{
\item{data}{A list or object to be converted to class \code{rcdf}.}
}
\value{
The input object with class set to \code{rcdf}.
}
\description{
Converts an existing list or compatible object into an object of class \code{rcdf}.
}
\examples{
my_list <- list(a = 1, b = 2)
rcdf_obj <- as_rcdf(my_list)
class(rcdf_obj)
}
