% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{decrypt_string}
\alias{decrypt_string}
\title{Decrypt string using RSA}
\usage{
decrypt_string(x, prv_key, password = NULL)
}
\arguments{
\item{x}{Encrypted base64-encoded string}

\item{prv_key}{A private key object or .pem file}

\item{password}{Passwor of the private key}
}
\value{
A decrpyted character of length 1
}
\description{
Decrypt string using RSA
}
\examples{
dir <- system.file("extdata", package = "rcdf")
pub_key <- file.path(dir, 'sample-public-key.pem')
prv_key <- file.path(dir, 'sample-private-key.pem')
x <- encrypt_string('hello', pub_key)
decrypt_string(x, prv_key = prv_key, password = '1234')
}
