% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{encrypt_string}
\alias{encrypt_string}
\title{Encrypt string using RSA}
\usage{
encrypt_string(x, pub_key)
}
\arguments{
\item{x}{A character of length 1}

\item{pub_key}{A public key object or .pem file}
}
\value{
Encrypted base64-encoded string
}
\description{
Encrypt string using RSA
}
