% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{generate_pw}
\alias{generate_pw}
\title{Generate a random password}
\usage{
generate_pw(length = 16, special_chr = TRUE)
}
\arguments{
\item{length}{Integer. The length of the password to generate. Default is \code{16}.}

\item{special_chr}{Logical. Whether to include special characters
(e.g., `!`, `@`, `#`, etc.) in the password. Default is \code{TRUE}.}
}
\value{
A character string representing the generated password.
}
\description{
This function generates a random password of a specified length. It includes
alphanumeric characters by default and can optionally include special characters.
}
\examples{
generate_pw()
generate_pw(32)
generate_pw(12, special_chr = FALSE)
}
