% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{get_attr}
\alias{get_attr}
\title{Get metadata attribute from RCDF data}
\usage{
get_attr(rcdf, attr)
}
\arguments{
\item{rcdf}{RCDF data}

\item{attr}{Valid metadata key.}
}
\value{
RCDF attribute/s or NULL
}
\description{
Get metadata attribute from RCDF data
}
\examples{
\dontrun{
# Assuming `df` is a valid RCDF object

get_attr(df, "area_names")

# To get nested attributes
get_attr(df, "meta.source_note")
}
}
