% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcdf-package.R
\docType{package}
\name{rcdf-package}
\alias{rcdf}
\alias{rcdf-package}
\title{rcdf: A Comprehensive Toolkit for Working with Encrypted Parquet Files}
\description{
Utilities for reading, writing, and managing RCDF files, including encryption and decryption support. It offers a flexible interface for handling data stored in encrypted Parquet format, along with metadata extraction, key management, and secure operations using AES and RSA encryptions.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://yng-me.github.io/rcdf/}
  \item \url{https://github.com/yng-me/rcdf}
  \item Report bugs at \url{https://github.com/yng-me/rcdf/issues}
}

}
\author{
\strong{Maintainer}: Bhas Abdulsamad \email{aeabdulsamad@gmail.com} (\href{https://orcid.org/0009-0002-5891-8124}{ORCID}) [copyright holder]

}
\keyword{internal}
