% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rcdf_list}
\alias{rcdf_list}
\title{Create an empty `rcdf` object}
\usage{
rcdf_list()
}
\value{
A list object of class `"rcdf"`.
}
\description{
Initializes and returns an empty `rcdf` object. This is a convenient constructor
for creating a new `rcdf`-class list structure.
}
\examples{
rcdf <- rcdf_list()
class(rcdf)
}
