% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_parquet.R
\name{read_parquet_as_db}
\alias{read_parquet_as_db}
\title{Read Parquet file as database}
\usage{
read_parquet_as_db(path, decryption_key, table_name = NULL, columns = NULL)
}
\arguments{
\item{path}{The file path to the Parquet file.}

\item{decryption_key}{A list containing \code{aes_key} and \code{aes_iv}. If provided, the Parquet file will be decrypted using these keys. Default is \code{NULL}.}

\item{table_name}{Database table name. If \code{NULL}, file name will be used as table name.}

\item{columns}{A character vector matching the column names available in the Parquet file.}
}
\value{
Lazy table from DuckDB connection
}
\description{
This function reads a Parquet file, optionally decrypting it using the provided decryption key. If no decryption key is provided, it reads the file normally without decryption. It supports reading Parquet files as Arrow tables or regular data frames, depending on the \code{as_arrow_table} argument.
}
\examples{
\dontrun{
# Using sample Parquet files from `mtcars` dataset
dir <- system.file("extdata", package = "rcdf")

# Encrypted
read_parquet_db(
  file.path(dir, "mtcars-encrypted.parquet"),
  decryption_key = 'rppqM5CuEqotys4wQq/g7xh6wpIjRozcAIbI9sagwKE='
)
}
}
