% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_rcdf_as.R
\name{write_rcdf_sqlite}
\alias{write_rcdf_sqlite}
\title{Title}
\usage{
write_rcdf_sqlite(data, path, db_name = "cbms_data", ..., parent_dir = NULL)
}
\arguments{
\item{parent_dir}{}
}
\description{
Title
}
