% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{collect}
\alias{collect}
\title{Collect}
\usage{
collect(data, ...)
}
\arguments{
\item{data}{A lazy data frame (e.g. from dbplyr or dtplyr) from database connection.}

\item{...}{Optional arguments}
}
\value{
A data frame
}
\description{
Collect
}
