% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{decrypt_string}
\alias{decrypt_string}
\title{Decrypt string using RSA}
\usage{
decrypt_string(x, prv_key, password = NULL)
}
\arguments{
\item{x}{Encrypted base64-encoded string}

\item{prv_key}{A private key object or .pem file}

\item{password}{Passwor of the private key}
}
\value{
A decrpyted character of length 1
}
\description{
Decrypt string using RSA
}
