% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_parquet.R
\name{read_parquet}
\alias{read_parquet}
\title{Read Parquet File with Optional Decryption}
\usage{
read_parquet(path, ..., decryption_key = NULL, as_arrow_table = TRUE)
}
\arguments{
\item{path}{The file path to the Parquet file.}

\item{...}{Additional arguments passed to `arrow::open_dataset()` when no decryption key is provided.}

\item{decryption_key}{A list containing `aes_key` and `aes_iv`. If provided, the Parquet file will be decrypted using these keys. Default is `NULL`.}

\item{as_arrow_table}{Logical. If `TRUE`, the function will return the result as an Arrow table. If `FALSE`, a regular data frame will be returned. Default is `TRUE`.}
}
\value{
An Arrow table or a data frame, depending on the value of `as_arrow_table`.
}
\description{
This function reads a Parquet file, optionally decrypting it using the provided decryption key. If no decryption key is provided, it reads the file normally without decryption. It supports reading Parquet files as Arrow tables or regular data frames, depending on the `as_arrow_table` argument.
}
\examples{
# Without decryption
df <- read_parquet("data.parquet")

# With decryption
decryption_key <- list(aes_key = "your_aes_key", aes_iv = "your_aes_iv")
df <- read_parquet("data_encrypted.parquet", decryption_key = decryption_key)
}
